+++ 3rd_party/darknet/src/blas.c
@@ -33,7 +33,7 @@ void variance_cpu(float *x, float *mean, int batch, int filters, int spatial, fl
         for(j = 0; j < batch; ++j){
             for(k = 0; k < spatial; ++k){
                 int index = j*filters*spatial + i*spatial + k;
-                variance[i] += pow((x[index] - mean[i]), 2);
+                variance[i] += powf((x[index] - mean[i]), 2);
             }
         }
         variance[i] *= scale;
@@ -47,7 +47,7 @@ void normalize_cpu(float *x, float *mean, float *variance, int batch, int filter
         for(f = 0; f < filters; ++f){
             for(i = 0; i < spatial; ++i){
                 int index = b*filters*spatial + f*spatial + i;
-                x[index] = (x[index] - mean[f])/(sqrt(variance[f]) + .000001f);
+                x[index] = (x[index] - mean[f])/(sqrtf(variance[f]) + .000001f);
             }
         }
     }
@@ -68,7 +68,7 @@ void mul_cpu(int N, float *X, int INCX, float *Y, int INCY)
 void pow_cpu(int N, float ALPHA, float *X, int INCX, float *Y, int INCY)
 {
     int i;
-    for(i = 0; i < N; ++i) Y[i*INCY] = pow(X[i*INCX], ALPHA);
+    for(i = 0; i < N; ++i) Y[i*INCY] = powf(X[i*INCX], ALPHA);
 }
 
 void axpy_cpu(int N, float ALPHA, float *X, int INCX, float *Y, int INCY)
